help type construction

+++ src/visualize/text.jl
@@ -29,7 +29,7 @@ end
 function visualize(text::AbstractString, s::Style, customizations=visualize_default(text, s))
     glyphs      = GPUVector(texture_buffer(process_for_gl(text)))
     positions   = GPUVector(texture_buffer(calc_position(glyphs)))
-    style_index = GPUVector(texture_buffer(fill(GLSpriteStyle(0,0), length(text))))
+    style_index = GPUVector(texture_buffer(fill(GLSpriteStyle(Uint16(0), Uint16(0)), length(text))))
     visualize(glyphs, positions, style_index, customizations[:model], s, customizations)  
 end
